Minimal changes on views
