+++ src/app/home/home.component.scss
@@ -5,3 +5,7 @@
 h3 {
     font-weight: normal;
 }
+
+.img-fluid-home {
+    max-height: 126px;
+}
