+++ src/app/products/new-product/new-product.component.html
@@ -1,25 +1,3 @@
-<ul class="nav d-flex align-items-center shadow p-2 mb-5">
-    <app-sidedraw class="p-2"></app-sidedraw>&nbsp;&nbsp;
-    <h3 class="title">Menú Principal</h3>
-    <hr class="mb-2" />
-    <div class="ms-auto p-2">
-        <div class="btn-group">
-            <button class="btn btn-info dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="true">
-        <!-- {{ user }} -->
-      </button>
-            <ul class="dropdown-menu">
-                <li><a href="#" class="dropdown-item">Ajustes</a></li>
-                <li><a href="#" class="dropdown-item">Ajustes</a></li>
-                <li><a href="#" class="dropdown-item">Ajustes</a></li>
-                <li>
-                    <hr class="dropdown-divider" />
-                </li>
-                <li><a href="#" class="dropdown-item"> Cerrar sesión</a></li>
-            </ul>
-        </div>
-    </div>
-</ul>
-
 <div class="container align-content-center">
     <div class="card shadow">
         <div class="card-header text-center">
