+++ src/app/providers/new-provider/new-provider.component.html
@@ -1,3 +1,4 @@
+<app-header></app-header>
 <div class="container">
     <div class="row mt-5">
         <div class="col-lg-8 offset-lg-2 col-md-10 offset-md-1 col-sm-12">
