+++ src/app/shared/components/header/header.component.html
@@ -1,1 +1,21 @@
-<p>header works!</p>
+<ul class="nav d-flex align-items-center shadow p-2 mb-5">
+    <app-sidedraw class="p-2"></app-sidedraw>&nbsp;&nbsp;
+    <h3 class="title">
+        {{tab}}
+    </h3>
+    <hr class="mb-2">
+    <div class="ms-auto p-2">
+        <div class="btn-group">
+            <button class="btn btn-info dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="true">{{user}}</button>
+            <ul class="dropdown-menu">
+                <li><a href="#" class="dropdown-item">Ajustes</a></li>
+                <li><a href="#" class="dropdown-item">Ajustes</a></li>
+                <li><a href="#" class="dropdown-item">Ajustes</a></li>
+                <li>
+                    <hr class="dropdown-divider">
+                </li>
+                <li><a href="#" class="dropdown-item"> Cerrar sesión</a></li>
+            </ul>
+        </div>
+    </div>
+</ul>
