+++ src/app/shared/components/sidedraw/sidedraw.component.scss
@@ -1,6 +1,9 @@
-.offcanvas-header,
+.offcanvas-header {
+    background-color: #ffe46c;
+}
+
 .offcanvas-body {
-    background: #ffee58;
+    background: linear-gradient(0deg, #ffc107, #ffc621, #fecc31, #fed13e, #fed64b, #feda56, #ffdf61, #ffe46c);
     // background: linear-gradient(0deg, rgba(189, 28, 53, 1) 0%, rgba(194, 21, 56, 1) 100%);
 }
